Only add defaults once
I have yet to encounter this, but I saw a potential problem with the way default Todos are handled. This commit adds a settings box that keeps track of whether or not they've been added. In the future I will add a settings page where users can reset the application to defaults.